+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/Actions/ActionId/Common/Kiosk/Success/index.tsx
