+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/Actions/ActionId/layout/index.tsx
@@ -68,13 +68,15 @@ export const ActionIdLayout = async (props: ActionIdLayout) => {
               <Typography variant={TYPOGRAPHY.R4}>Proof debugging</Typography>
             </Tab>
 
-            <Tab
-              className="py-4"
-              href={`/teams/${params!.teamId}/apps/${params!.appId}/actions/${params!.actionId}/kiosk`}
-              segment={"kiosk"}
-            >
-              <Typography variant={TYPOGRAPHY.R4}>Kiosk</Typography>
-            </Tab>
+            {!isOnChainApp && (
+              <Tab
+                className="py-4"
+                href={`/teams/${params!.teamId}/apps/${params!.appId}/actions/${params!.actionId}/kiosk`}
+                segment={"kiosk"}
+              >
+                <Typography variant={TYPOGRAPHY.R4}>Kiosk</Typography>
+              </Tab>
+            )}
 
             {isEnoughPermissions && (
               <Tab
